Changed Travis language to Python
This may seem like an odd choice, but it is required since Camlann
cannot be compiled using a simple `make` as most C project could, and as
TravisCI assumes it can be, This is a direct and inherent result of it's
purpose. The idea here is to get Travis to just run a specific script
(camlann_CI.mk) and use that to determine "build" success/failure.